abstract out context creation in tests
